correctly handle self-addr port 0
